Compare-BERT-Models/requirements.txt
altair==4.2.0
anyio==3.6.1
attrs==22.2.0
Automat==20.2.0
bcrypt==3.2.0
blinker==1.4
blis==0.7.9
boto3==1.26.113
botocore==1.29.113
Brlapi==0.8.4
catalogue==2.0.8
certifi==2022.12.7
chardet==4.0.0
charset-normalizer==3.0.1
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.5
command-not-found==0.3
commonmark==0.9.1
confection==0.0.4
configobj==5.0.6
constantly==15.1.0
contourpy==1.0.6
corenlp==0.0.14
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.6
distro==1.7.0
distro-info==1.1+ubuntu1
entrypoints==0.4
filelock==3.9.0
fonttools==4.38.0
fst-pso==1.8.1
FuzzyTM==2.0.5
gensim==4.3.0
h11==0.13.0
httpcore==0.15.0
httplib2==0.20.4
httpx==0.23.0
hyperlink==21.0.0
idna==3.4
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
keyring==23.8.2
kiwisolver==1.4.4
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
louis==3.22.0
MarkupSafe==2.1.1
matplotlib==3.6.3
miniful==0.0.6
more-itertools==8.10.0
murmurhash==1.0.9
nala==0.11.1
netifaces==0.11.0
nltk==3.8.1
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.0
olefile==0.46
packaging==23.0
pandas==1.5.2
pathy==0.10.1
pexpect==4.8.0
Pillow==9.2.0
platformdirs==3.0.0
preshed==3.0.8
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycups==2.0.1
pydantic==1.10.4
pyFUME==0.2.25
Pygments==2.12.0
PyGObject==3.42.2
PyHamcrest==2.0.2
PyJWT==2.4.0
pyOpenSSL==21.0.0
pyparsing==3.0.7
pyrsistent==0.19.3
python-apt==2.4.0+22.10
python-dateutil==2.8.1
python-debian==0.1.46
pytorch-transformers==1.2.0
pytz==2022.7
pyxdg==0.27
PyYAML==5.4.1
regex==2022.10.31
reportlab==3.6.11
requests==2.28.2
rfc3986==1.5.0
rich==12.4.4
s3transfer==0.6.0
sacremoses==0.0.53
scikit-learn==1.2.0
scipy==1.10.0
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==3.3.3
sentencepiece==0.1.98
service-identity==18.1.0
simpful==2.9.0
six==1.16.0
smart-open==6.3.0
sniffio==1.2.0
socksio==1.0.0
spacy==3.4.4
spacy-legacy==3.0.11
spacy-loggers==1.0.4
srsly==2.4.5
ssh-import-id==5.11
systemd-python==235
thinc==8.1.6
threadpoolctl==3.1.0
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchaudio==0.13.1
torchtext==0.14.1
torchvision==0.14.1
tqdm==4.64.1
Twisted==22.4.0
typer==0.7.0
typing_extensions==4.4.0
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.14
virtualenv==20.19.0
wadllib==1.3.6
wasabi==0.10.1
xdg==5
xgboost==1.7.3
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
